chore: split out code into more modules
