+++ src/constants.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const LOOP_SENTINEL = 1_000_000
+
+module.exports = {
+    LOOP_SENTINEL
+}
